Matrice : (int *) -> (signed char *)

+++ projet_lem_in/srcs_lem_in/li_bellman_ford.c
@@ -29,7 +29,7 @@ void	aux_li_bellman_ford(t_data *data)
 			if (*((*(data->map + i))->pipes + target) != 0 && i != target)
 			{
 				wei = (*(data->map + i))->weight
-					+ *((*(data->map + i))->pipes + target);
+					+ (int)(*((*(data->map + i))->pipes + target));
 				if (wei < (*(data->map + target))->weight)
 				{
 					(*(data->map + target))->ancestor = i;
